LC-distinct-subsequences [Runtime: N/A; Memory: N/A]

+++ string/distinct-subsequences.cpp
@@ -20,6 +20,6 @@ public:
             }
         }
 
-        return dp[n][m];
+        return (int)dp[n][m];
     }
 };
